+++ providers/firebase-rest.js
