+++ providers/supabase-admin.js
@@ -1,0 +1,8 @@
+import {createClient} from '@supabase/supabase-js'
+import config from 'utils/config'
+
+const {SUPABASE_URL, SUPABASE_ANON_KEY} = config
+
+const adminClient = createClient(SUPABASE_URL, SUPABASE_ANON_KEY)
+
+export default adminClient
